Forum

+++ View/subjectView.php
@@ -4,30 +4,53 @@
         foreach ($var['subject'] as $subject) { ?>
             <h1 class="center"><?=$subject->getTitle()?></h1>
             <p class="center gray">Par <?=$subject->getUserFk()->getPseudo()?></p>
-            <div class="height">
-                <a href="../index.php?controller=subjects&action=update&id=<?=$subject->getId()?>&id2=<?=$subject->getCategorieFk()->getId()?>" class="button buttonAbsolute1"><i class="fas fa-edit"></i></a>
-            </div>
-            <div class="height">
-                <a href="#" class="button buttonAbsolute1"><i class="fas fa-archive"></i></a>
-            </div>
-            <div class="height">
-                <a href="../index.php?controller=subjects&action=delete&id=<?=$subject->getId()?>" class="button buttonAbsolute1"><i class="fas fa-trash-alt"></i></a>
-            </div>
+            <?php
+            if (isset($_SESSION['id'])) {
+                // Admin and moderator can edit and the user who created the topic too
+                if ($_SESSION['role_fk'] !== "2" || $subject->getUserFk()->getId() == $_SESSION['id']) { ?>
+                    <div class="height">
+                        <a href="../index.php?controller=subjects&action=update&id=<?=$subject->getId()?>&id2=<?=$subject->getCategorieFk()->getId()?>" class="button buttonAbsolute1"><i class="fas fa-edit"></i></a>
+                    </div>
+                <?php
+                }
+                // That the admin and the moderator who can archive
+                if ($_SESSION['role_fk'] !== "2") { ?>
+                    <div class="height">
+                        <a href="#" class="button buttonAbsolute1"><i class="fas fa-archive"></i></a>
+                    </div>
+                <?php
+                }
+                // That the admin who can delete
+                if ($_SESSION['role_fk'] === "1") { ?>
+                    <div class="height">
+                        <a href="../index.php?controller=subjects&action=delete&id=<?=$subject->getId()?>" class="button buttonAbsolute1"><i class="fas fa-trash-alt"></i></a>
+                    </div>
+                <?php
+                }
+            }
+                ?>
             <div class="flexCenter">
-                <img class="imageCategorie" src="<?=$subject->getPicture()?>">
+                <img class="imageCategorie" src="<?=$subject->getPicture()?>" alt="<?=$subject->getTitle()?>">
             </div>
             <h2 class="center"><?=$subject->getDescription()?></h2>
             <p class="text"><?=$subject->getText()?></p>
+
+            <div class="lineHorizontal"></div>
+            <div class="flexRow flexCenter">
+                <h2 class="gray">Commentaires</h2>
+                <?php
+                if (isset($_SESSION['id'])) {?>
+                    <a href="../index.php?controller=comments&action=new&id=<?=$_SESSION['id']?>&id2=<?=$subject->getId()?>&id3=<?=$subject->getCategorieFk()->getId()?>" class="button buttonAbsolute1"><i class="fas fa-plus"></i> Ajouter un commentaire</a>
+                    <?php
+                }
+                    ?>
+            </div>
         <?php
         }
     }
     ?>
 
-    <div class="lineHorizontal"></div>
-    <div class="flexRow flexCenter">
-        <h2 class="gray">Commentaires</h2>
-        <a href="../index.php?controller=comments&action=new" class="button buttonAbsolute1"><i class="fas fa-plus"></i> Ajouter un commentaire</a>
-    </div>
+
 
     <?php
     if (isset($var['comments'])) {
@@ -35,9 +58,24 @@
             <div class="comments flexColumn">
                 <div class="pseudo gray flexRow align">
                     <p><?=$comment->getUserFk()->getPseudo()?> / <?=$comment->getDate()?></p>
-                    <a href="Update/updateCommentView.php" class="button3"><i class="fas fa-edit"></i></a>
-                    <a href="Delete/deleteCommentView.php" class="button3 buttonPos3"><i class="fas fa-trash-alt"></i></a>
-                    <a href="#" class="button3 buttonPos4"><i class="fas fa-archive"></i></a>
+                    <?php
+                    if (isset($_SESSION['id'])) {
+                        // Admin and moderator can edit and the user who created the topic too
+                        if ($_SESSION['role_fk'] !== "2" || $comment->getUserFk()->getId() == $_SESSION['id']) { ?>
+                            <a href="../index.php?controller=comments&action=update&id=<?=$comment->getId()?>" class="button3"><i class="fas fa-edit"></i></a>
+                        <?php
+                        }
+                        // That the admin who can delete
+                        if ($_SESSION['role_fk'] === "1") { ?>
+                            <a href="../index.php?controller=comments&action=delete&id=<?=$comment->getId()?>" class="button3 buttonPos3"><i class="fas fa-trash-alt"></i></a>
+                        <?php
+                        }
+                        // That the admin and the moderator who can archive
+                        if ($_SESSION['role_fk'] !== "2") { ?>
+                            <a href="#" class="button3 buttonPos4"><i class="fas fa-archive"></i></a>
+                        <?php
+                        }
+                    }?>
                 </div>
                 <div class="comment">
                     <p><?=$comment->getComment()?></p>
